+++ R/anova.logbinlist.r
@@ -1,3 +1,24 @@
+#  Based on File src/library/stats/R/glm.R
+#  Part of the R package, https://www.R-project.org
+#
+#  Modified by Mark W. Donoghoe
+#     28/08/2015 - to work with logbin objects
+#
+#  Copyright (C) 1995-2020 The R Core Team
+#
+#  This program is free software; you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation; either version 2 of the License, or
+#  (at your option) any later version.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  A copy of the GNU General Public License is available at
+#  https://www.R-project.org/Licenses/
+
 anova.logbinlist <- function(object, ..., test = NULL) {
   doscore <- !is.null(test) && test == "Rao"
   responses <- as.character(lapply(object, function(x) {deparse(formula(x)[[2L]])}))